LPS-180588 Extract VisibilityButton component to a separate file
